Lab-4 done

+++ Lab-4/app/Http/Controllers/TournamentController.php
@@ -6,6 +6,9 @@ use App\Models\Tournament;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use function Laravel\Prompts\select;
+use App\Models\User;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
 
 class TournamentController extends Controller
 {
@@ -19,7 +22,8 @@ class TournamentController extends Controller
         $this->tournaments = Tournament::all();
     }
 
-    function debug_to_console($data) {
+    function debug_to_console($data)
+    {
         $output = $data;
         if (is_array($output))
             $output = implode(',', $output);
@@ -33,16 +37,22 @@ class TournamentController extends Controller
     }
     public function create()
     {
+        $user = Auth::user();
         $this->debug_to_console("Create");
-        return view('tournaments.create');
+        if (Gate::forUser($user)->allows('create-tournament')) {
+            return view('tournaments.create');
+        } else {
+            abort(403);
+        }
     }
     public function store(Request $request)
     {
         $request->validate([
             'fullName' => 'required|max:255',
         ]);
-
-        $tournament = Tournament::create($request->all());
+        
+        $user = Auth::user();
+        Tournament::create(array_merge($request->all(), ['creator_user_id' => $user->id]));
 
         return redirect('/tournaments/')->with('success', 'Tournament created successfully!');
     }
@@ -51,14 +61,19 @@ class TournamentController extends Controller
         $this->debug_to_console("Show");
         $tournament = Tournament::findOrFail($id);
         return view('tournaments.show', ['tournament' => $tournament]);
-        //
     }
     public function edit($id)
     {
+        $user = Auth::user();
         $this->debug_to_console("Edit");
         $tournament = Tournament::findOrFail($id);
-        return view('tournaments.edit', ['tournament' => $tournament]);
-        //
+        $toConsole = 'User ID: ' . (string) $user->id . ' Tournament creator ID: ' . (string) $tournament->creator_user_id;
+        $this->debug_to_console($toConsole);
+        if (Gate::forUser($user)->allows('edit-tournament', $tournament)) {
+            return view('tournaments.edit', ['tournament' => $tournament]);
+        } else {
+            abort(403);
+        }
     }
     public function update(Request $request, $id)
     {
@@ -72,8 +87,13 @@ class TournamentController extends Controller
     }
     public function destroy($id)
     {
+        $user = Auth::user();
         $tournament = Tournament::findOrFail($id);
-        $tournament->delete();
-        return redirect('/tournaments')->with('success', 'Tournament deleted successfully!');
+        if (Gate::forUser($user)->allows('delete-tournament', $tournament)) {
+            $tournament->delete();
+            return redirect('/tournaments')->with('success', 'Tournament deleted successfully!');
+        } else {
+            abort(403);
+        }
     }
 }
